Unsigned comparison

+++ Ryujinx.HLE/OsHle/Process.cs
@@ -327,7 +327,7 @@ namespace Ryujinx.HLE.OsHle
 
             for (int Index = Executables.Count - 1; Index >= 0; Index--)
             {
-                if (Position >= Executables[Index].ImageBase)
+                if ((ulong)Position >= (ulong)Executables[Index].ImageBase)
                 {
                     return Executables[Index];
                 }
